LPS-53274 Move com.liferay.polls.lar to com.liferay.polls.lar.test
