+++ modules/apps/polls/polls-test/test/integration/com/liferay/polls/lar/test/PollsPortletDataHandlerTest.java
@@ -12,9 +12,10 @@
  * details.
  */
 
-package com.liferay.polls.lar;
+package com.liferay.polls.lar.test;
 
 import com.liferay.polls.constants.PollsPortletKeys;
+import com.liferay.polls.lar.PollsPortletDataHandler;
 import com.liferay.polls.model.PollsChoice;
 import com.liferay.polls.model.PollsQuestion;
 import com.liferay.polls.util.test.PollsTestUtil;
